+++ SecurityFever.Library/Properties/AssemblyVersion.cs
@@ -1,6 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.InteropServices;
-
-// Version of the product.
-[assembly: AssemblyVersion("3.0.0")]
-[assembly: AssemblyFileVersion("3.0.0")]